requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-groq>=0.1.0
langchain-chroma>=0.0.3
chromadb>=0.4.18
pydantic>=2.5.2
numpy>=1.24.0
flask>=2.3.0
python-dotenv>=1.0.0